Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: 26ade7cae88d84ef9e1aed95b3df2cd6d5f893ee [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6] [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 26ade7cae88d84ef9e1aed95b3df2cd6d5f893ee [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6] [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly aa991c922354f0a67fac08fee456ca60a30de968 [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]]]
Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af
Former-commit-id: d395fc27fc163b2f1f124017e47c57f63e8859d1 [formerly 5977b049c1eda298ce206a8d5fc34b8f235639c5] [formerly 53a85ec6b9da7b5e37777469d995ba034e153ce6 [formerly 10ea02b9366284a7b97f57ffdb7ac54688cfa0c4]]
Former-commit-id: 165edd61e4cf5f8b3c8f4d5abc9d215a41539cc6 [formerly 053cf978a7bcf69a7c3cc61bc2be8386791b90aa]
Former-commit-id: 8a256eb22803e88cf9bc9e45e778f1aa2bdb224f

+++ edexOsgi/com.raytheon.edex.plugin.ldadmanual/src/com/raytheon/edex/plugin/ldadmanual/dao/ManualLdadRecord.java
@@ -62,6 +62,7 @@ import com.vividsolutions.jts.geom.Geometry;
  *                                    PluginDataObject.
  * Aug 30, 2013  2298     rjpeter     Make getPluginName abstract
  * Jun 11, 2014  2061     bsteffen    Remove IDecoderGettable
+ * Jul 23, 2014  3410     bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -2102,7 +2103,7 @@ public class ManualLdadRecord extends PluginDataObject implements
      * 
      * @return The geometry latitude.
      */
-    public Double getLatitude() {
+    public Float getLatitude() {
         return location.getLatitude();
     }
 
@@ -2111,7 +2112,7 @@ public class ManualLdadRecord extends PluginDataObject implements
      * 
      * @return The geometry longitude.
      */
-    public Double getLongitude() {
+    public Float getLongitude() {
         return location.getLongitude();
     }
